플레이어 장비 장착, 아이템 로직
플레이어 아이템, 아이템 SO, 아이템 로직SO, 추가 (오브젝트 풀 사용)
EquipSlot 로직 변경 (장착 해제 사용 로직 배열 추가)

+++ Assets/03_Scripts/UI/HealthManager.cs
@@ -16,7 +16,7 @@ public class HealthManager : MonoBehaviour
         if (m_pPlayerHealth.CurrentHP == 0)
             fHPRatio = 0.0f;
 
-        fHPRatio = m_pPlayerHealth.CurrentHP / m_pPlayerHealth.MaxHP;
+        fHPRatio = (float)m_pPlayerHealth.CurrentHP / m_pPlayerHealth.MaxHP;
 
         if(fHPRatio == 0.0f)
             m_pHealthImage.fillAmount = 0;
